requirements.txt
lancedb>=0.10.0
pyarrow>=14.0.0
numpy>=1.24.0
openai>=1.40.0
python-dotenv>=1.0.1
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
joblib>=1.3.2
tqdm>=4.66.0
fastapi>=0.109.0
uvicorn>=0.27.0
docling>=1.0.0
python-multipart>=0.0.9
reportlab>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
rapidfuzz>=3.0.0
pydantic>=2.0.0
